admin can manage parks, albeit with a single field

+++ db/migrate/20121025015408_create_parks.rb
@@ -6,7 +6,7 @@ class CreateParks < ActiveRecord::Migration
       t.float :latitude
       t.float :longitude
       t.text :contact_info
-      t.decimal :park_size
+      t.float :park_size
       t.text :address
       t.text :vimeo_embed
       t.string :flicker_pool
